Service Layer Revised
I added files for the Service Layer, which connects the database to the Client Layer; however, I was unable to execute the files properly so that the data in the database appears on the browser. I also was unsuccessful in implementing CRUD features where the user creates an account by registration, reads the information that was provided during registration, updates the account, and deletes the account.